Fix project strcuture
Migrate RDSeffSwitch to tycho.
